+++ tests/Support/Attributes/CustomTag.php
@@ -5,11 +5,11 @@ declare(strict_types=1);
 namespace JeroenG\Autowire\Tests\Support\Attributes;
 
 use Attribute;
-use JeroenG\Autowire\Attribute\AutotagInterface;
+use JeroenG\Autowire\Attribute\TagInterface;
 use ReflectionClass;
 
 #[Attribute(Attribute::TARGET_CLASS)]
-class CustomAutotag implements AutotagInterface
+class CustomTag implements TagInterface
 {
     public function getTag(ReflectionClass $targetInterface): string
     {