+++ tests/Support/Subject/Contracts/GoodafternoonInterface.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace JeroenG\Autowire\Tests\Support\Subject\Contracts;
 
-use JeroenG\Autowire\Attribute\Autotag;
+use JeroenG\Autowire\Attribute\Tag;
 
-#[Autotag]
+#[Tag]
 interface GoodafternoonInterface
 {
     public function goodafternoon(): string;