+++ tests/Support/Subject/Contracts/GoodeveningInterface.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace JeroenG\Autowire\Tests\Support\Subject\Contracts;
 
-use JeroenG\Autowire\Attribute\Autotag;
+use JeroenG\Autowire\Attribute\Tag;
 
-#[Autotag('evening')]
+#[Tag('evening')]
 interface GoodeveningInterface
 {
     public function goodevening(): string;