+++ tests/Support/Subject/Contracts/GoodmorningInterface.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace JeroenG\Autowire\Tests\Support\Subject\Contracts;
 
-use JeroenG\Autowire\Tests\Support\Attributes\CustomAutotag;
+use JeroenG\Autowire\Tests\Support\Attributes\CustomTag;
 
-#[CustomAutotag]
+#[CustomTag]
 interface GoodmorningInterface
 {
     public function goodmorning(): string;