+++ changelog.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ## [1.6]
-- New `Autotag` attribute for automatically tagging implementations of an interface
+- New `Tag` attribute for automatically tagging implementations of an interface
 
 ## [1.5]
 - New `Listen` attribute for autowiring events to event listeners